refactor: Implement classroom reducer
- move `classrooms` field from chat reducer initial state to `classroomSlice.js`.
- move `joinClassroom`, `leaveClassroom` and `getClassrooms` thunks, along with extra reducers state handling from `chatSlice.js` to `classroomSlice.js`.
- move `joinClassroom`, `leaveClassroom`, and `getClassrooms` from `chatService.js` to `classroomService.js`.
- move `handleError` function from `chatService.js`. to a new module `service.errorHandler.js`.
- integrate `handleError` in `classroomService.js`.
- integrate `classroomReducer` into `store` in `store.js`.

Note: this commit adds code readability and maintenance by focusing on modularity and separation of concerns.
