+++ frontend/src/features/chat/chatService.js
@@ -1,4 +1,5 @@
 import axios from "axios";
+import handleError from "../service.errorHandler.js";
 
 const API_URL = "http://127.0.0.1:8000/api/chat/";
 
@@ -18,40 +19,6 @@ const sendFriendMessage = async (messageData, token) => {
   }
 };
 
-const joinClassroom = async (classroomId, token) => {
-  try {
-    const response = await axios.post(
-      API_URL + "classroom/join",
-      { classroomId },
-      { headers: { Authorization: `Bearer ${token}` } }
-    );
-    return response.data;
-  } catch (error) {
-    handleError(error);
-    throw error;
-  }
-};
-
-const leaveClassroom = async (classroomId, token) => {
-  try {
-    const config = {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    };
-
-    const response = await axios.post(
-      `${API_URL}classroom/leave`,
-      { classroomId },
-      config
-    );
-    return response.data;
-  } catch (error) {
-    handleError(error);
-    throw error;
-  }
-};
-
 const getUserMessages = async (token) => {
   try {
     const config = {
@@ -61,19 +28,6 @@ const getUserMessages = async (token) => {
     };
 
     const response = await axios.get(API_URL, config);
-    return response.data;
-  } catch (error) {
-    handleError(error);
-    throw error;
-  }
-};
-
-const getClassrooms = async (token) => {
-  try {
-    const response = await axios.get(API_URL + "classrooms", {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-
     return response.data;
   } catch (error) {
     handleError(error);
@@ -100,23 +54,10 @@ const getClassroomMessages = async (classroomId, token) => {
   }
 };
 
-const handleError = (error) => {
-  if (error.response) {
-    console.error("Server Error:", error.response.data);
-  } else if (error.request) {
-    console.error("No response received:", error.request);
-  } else {
-    console.error("Error setting up request:", error.message);
-  }
-};
-
 const chatService = {
   sendFriendMessage,
-  joinClassroom,
   getUserMessages,
   getClassroomMessages,
-  getClassrooms,
-  leaveClassroom,
 };
 
 export default chatService;
