+++ frontend/src/features/chat/chatSlice.js
@@ -6,7 +6,6 @@ const initialState = {
   isSuccess: false,
   isError: false,
   errorMessage: "",
-  classrooms: [],
   messages: {},
 };
 
@@ -58,54 +57,6 @@ export const sendClassroomMessage = createAsyncThunk(
   }
 );
 
-export const joinClassroom = createAsyncThunk(
-  "chat/joinClassroom",
-  async (classroomId, thunkAPI) => {
-    try {
-      const { token } = thunkAPI.getState().auth.user;
-
-      if (!token) {
-        throw new Error("Token not found");
-      }
-
-      return await chatService.joinClassroom(classroomId, token);
-    } catch (error) {
-      const message =
-        (error.response &&
-          error.response.data &&
-          error.response.data.message) ||
-        error.message ||
-        error.toString();
-
-      return thunkAPI.rejectWithValue(message);
-    }
-  }
-);
-
-export const leaveClassroom = createAsyncThunk(
-  "chat/leaveClassroom",
-  async (classroomId, thunkAPI) => {
-    try {
-      const token = thunkAPI.getState().auth.user.token;
-
-      if (!token) {
-        throw new Error("Token not found");
-      }
-
-      return await chatService.leaveClassroom(classroomId, token);
-    } catch (error) {
-      const message =
-        (error.response &&
-          error.response.data &&
-          error.response.data.message) ||
-        error.message ||
-        error.toString();
-
-      return thunkAPI.rejectWithValue(message);
-    }
-  }
-);
-
 export const getUserMessages = createAsyncThunk(
   "chat/getUserMessages",
   async (_, thunkAPI) => {
@@ -131,31 +82,6 @@ export const getClassroomMessages = createAsyncThunk(
     try {
       const token = thunkAPI.getState().auth.user.token;
       return await chatService.getClassroomMessages(classroomId, token);
-    } catch (error) {
-      const message =
-        (error.response &&
-          error.response.data &&
-          error.response.data.message) ||
-        error.message ||
-        error.toString();
-
-      return thunkAPI.rejectWithValue(message);
-    }
-  }
-);
-
-export const getClassrooms = createAsyncThunk(
-  "chat/getClassrooms",
-  async (_, thunkAPI) => {
-    try {
-      const { token } = thunkAPI.getState().auth.user;
-      if (!token) {
-        throw new Error("Token not found");
-      }
-
-      const data = await chatService.getClassrooms(token);
-
-      return data;
     } catch (error) {
       const message =
         (error.response &&
@@ -215,23 +141,6 @@ const chatSlice = createSlice({
         state.isError = true;
         state.errorMessage = action.payload || "Failed to send message";
       })
-      .addCase(joinClassroom.pending, (state) => {
-        state.isLoading = true;
-        state.isError = false;
-        state.errorMessage = "";
-      })
-      .addCase(joinClassroom.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.isSuccess = true;
-        state.isError = false;
-        state.classrooms.push(action.payload);
-      })
-      .addCase(joinClassroom.rejected, (state, action) => {
-        state.isLoading = false;
-        state.isSuccess = false;
-        state.isError = true;
-        state.errorMessage = action.payload || "Failed to join classroom";
-      })
       .addCase(getUserMessages.pending, (state) => {
         state.isLoading = true;
         state.errorMessage = "";
@@ -275,38 +184,6 @@ const chatSlice = createSlice({
         state.isError = true;
         state.errorMessage =
           action.payload || "Failed to get classroom messages";
-      })
-      .addCase(getClassrooms.pending, (state) => {
-        state.isLoading = true;
-        state.isError = false;
-        state.errorMessage = "";
-      })
-      .addCase(getClassrooms.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.isSuccess = true;
-        state.isError = false;
-        state.errorMessage = "";
-        state.classrooms = action.payload;
-      })
-      .addCase(getClassrooms.rejected, (state, action) => {
-        state.isLoading = false;
-        state.isSuccess = false;
-        state.isError = true;
-        state.errorMessage =
-          action.payload.message || "Failed to load classrooms";
-      })
-      .addCase(leaveClassroom.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.isSuccess = true;
-        state.errorMessage = "";
-        state.classrooms = state.classrooms.filter(
-          (classroom) => classroom._id !== action.payload._id
-        );
-      })
-      .addCase(leaveClassroom.rejected, (state, action) => {
-        state.isSuccess = false;
-        state.isError = true;
-        state.errorMessage = action.payload || "Failed to leave classroom";
       });
   },
 });
